requirements.txt
bcrypt==4.1.3
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
setuptools==70.0.0
tenacity==8.3.0
Werkzeug==3.0.3
wheel==0.43.0
python-dotenv==1.0.1
APScheduler==3.10.4
